refactor : 주문 도메인 모듈 디렉토리 구조 수정
