+++ src/main/java/com/pintoss/auth/module/order/api/OrderController.java
@@ -3,8 +3,8 @@ package com.pintoss.auth.module.order.api;
 import com.pintoss.auth.common.dto.ApiResponse;
 import com.pintoss.auth.module.order.api.dto.CreateOrderRequest;
 import com.pintoss.auth.module.order.api.dto.CreateOrderResponse;
-import com.pintoss.auth.module.order.usecase.CancelOrderService;
-import com.pintoss.auth.module.order.usecase.CreateOrderService;
+import com.pintoss.auth.module.order.application.CancelOrderService;
+import com.pintoss.auth.module.order.application.CreateOrderService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.access.prepost.PreAuthorize;
