+++ src/main/java/com/pintoss/auth/module/order/api/dto/CreateOrderRequest.java
@@ -1,7 +1,7 @@
 package com.pintoss.auth.module.order.api.dto;
 
-import com.pintoss.auth.module.order.model.PaymentMethodType;
-import com.pintoss.auth.module.order.usecase.dto.OrderItemRequest;
+import com.pintoss.auth.module.order.application.model.PaymentMethodType;
+import com.pintoss.auth.module.order.application.dto.OrderItemRequest;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
