+++ src/main/java/com/pintoss/auth/module/order/application/flow/OrderFactory.java
@@ -1,0 +1,11 @@
+package com.pintoss.auth.module.order.application.flow;
+
+import com.pintoss.auth.module.order.application.model.Order;
+
+public class OrderFactory {
+
+    public Order create(){
+        return null;
+    }
+
+}
