+++ src/main/java/com/pintoss/auth/module/order/application/flow/OrderValidator.java
@@ -1,6 +1,6 @@
-package com.pintoss.auth.module.order.usecase.service;
+package com.pintoss.auth.module.order.application.flow;
 
-import com.pintoss.auth.module.order.usecase.dto.OrderItemRequest;
+import com.pintoss.auth.module.order.application.dto.OrderItemRequest;
 import com.pintoss.auth.module.voucher.model.Voucher;
 import java.util.List;
 import java.util.Map;
