+++ src/main/java/com/pintoss/auth/module/order/application/model/OrderItem.java
@@ -1,4 +1,4 @@
-package com.pintoss.auth.module.order.model;
+package com.pintoss.auth.module.order.application.model;
 
 import com.pintoss.auth.common.exception.client.BadRequestException;
 import jakarta.persistence.Column;
